Debug CORS error by adding CrossOrigin in controller

+++ quizApp/src/main/java/com/nucleusTeq/quizApp/repositories/QuestionRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface QuestionRepository extends JpaRepository<Questions,Integer> {
+public interface QuestionRepository extends JpaRepository<Questions,Long> {
 
     List<Questions> findByDifficultyLevelAndCategory_Id(String difficultyLevel, long categoryId);
 }
